Changes resulting from world_wiki 0.1.0
A few features are still disabled, like Notes and AttachmentsList, but
those will be back soon. I think there's still a little cruft in there.

Wiki links in all existing documents need to be updated to use the
double-bracket syntax. All <dm/> and <vp/> tags need to be updated
to use the squiggly syntax (see lib/convert.rb for help).